+++ app/helpers/redis_helper.rb
@@ -1,0 +1,12 @@
+module RedisHelper
+  def new_redis_client
+    conf_file = File.join('config', 'redis.yml')
+
+    if File.exists?(conf_file)
+      conf = YAML.load(File.read(conf_file))
+      conf[Rails.env.to_s].blank? ? Redis.new : Redis.new(conf[Rails.env.to_s])
+    else
+      Redis.new
+    end
+  end
+end
