+++ app/workers/feed_update_worker.rb
@@ -5,7 +5,7 @@ class FeedUpdateWorker
   recurrence { minutely.second_of_minute (0..60).step(5).to_a }
 
   def perform
-    HitFoxFeed.new(ENV['GITHUB_USER'],'HitFox',ENV['GITHUB']).update
+    HitFoxFeed.new.update
   end
 end
